Make the existing Telesales cancellation journey available to the public
https://eaflood.atlassian.net/browse/IWTF-5045

• Cancellation journey is available in Websales
• Cancellation journey is concealed behind a feature switch for Web sales.
• Feature switch is off in production
