+++ packages/gafl-webapp-service/src/routes/__tests__/routes.spec.js
@@ -1,3 +1,13 @@
+import {
+  CANCEL_RP_IDENTIFY,
+  CANCEL_RP_DETAILS,
+  CANCEL_RP_CONFIRM,
+  CANCEL_RP_COMPLETE,
+  CANCEL_RP_AGREEMENT_NOT_FOUND,
+  CANCEL_RP_LICENCE_NOT_FOUND,
+  CANCEL_RP_ALREADY_CANCELLED
+} from '../../uri.js'
+
 const mockErrorRoutes = [Symbol('error')]
 jest.mock('../error-routes.js', () => mockErrorRoutes)
 
@@ -6,6 +16,17 @@ jest.mock('../error-test-routes.js', () => mockErrorTestingRoutes)
 
 const mockTelesalesRoutes = [Symbol('telesales')]
 jest.mock('../telesales-routes.js', () => mockTelesalesRoutes)
+
+jest.mock('@defra-fish/connectors-lib')
+const getCancelRPURIs = () => [
+  CANCEL_RP_IDENTIFY.uri,
+  CANCEL_RP_DETAILS.uri,
+  CANCEL_RP_CONFIRM.uri,
+  CANCEL_RP_COMPLETE.uri,
+  CANCEL_RP_AGREEMENT_NOT_FOUND.uri,
+  CANCEL_RP_LICENCE_NOT_FOUND.uri,
+  CANCEL_RP_ALREADY_CANCELLED.uri
+]
 
 describe('route', () => {
   beforeEach(() => {
@@ -47,5 +68,37 @@ describe('route', () => {
     process.env.ERROR_PAGE_ROUTE = 'false'
     const routes = require('../routes.js')
     expect(routes.default).toEqual(expect.not.arrayContaining(mockErrorTestingRoutes))
+  })
+})
+
+describe('cancellation route journey behaves as expected', () => {
+  beforeEach(jest.clearAllMocks)
+
+  it('adds the cancellation route journey if SHOW_CANCELLATION_JOURNEY is set to true', () => {
+    process.env.SHOW_CANCELLATION_JOURNEY = 'true'
+    jest.isolateModules(() => {
+      const routesPaths = require('../routes.js').default.map(route => route.path)
+      expect(routesPaths).toEqual(expect.arrayContaining(getCancelRPURIs()))
+    })
+  })
+
+  it('omits the cancellation route journey if SHOW_CANCELLATION_JOURNEY is set to false', () => {
+    process.env.SHOW_CANCELLATION_JOURNEY = 'false'
+    jest.isolateModules(() => {
+      const routes = require('../routes.js').default
+      const cancelRPURIs = getCancelRPURIs()
+      const cancelRPRoutes = routes.filter(route => cancelRPURIs.includes(route.path))
+      expect(cancelRPRoutes).toHaveLength(0)
+    })
+  })
+
+  it('omits the cancellation route journey if SHOW_CANCELLATION_JOURNEY is not present', () => {
+    delete process.env.SHOW_CANCELLATION_JOURNEY
+    jest.isolateModules(() => {
+      const routes = require('../routes.js').default
+      const cancelRPURIs = getCancelRPURIs()
+      const cancelRPRoutes = routes.filter(route => cancelRPURIs.includes(route.path))
+      expect(cancelRPRoutes).toHaveLength(0)
+    })
   })
 })
