+++ packages/gafl-webapp-service/src/routes/__tests__/telesales-routes.spec.js
@@ -1,24 +1,7 @@
 import { setupEnvironment } from '../../__mocks__/openid-client.js'
-import {
-  CANCEL_RP_IDENTIFY,
-  CANCEL_RP_DETAILS,
-  CANCEL_RP_CONFIRM,
-  CANCEL_RP_COMPLETE,
-  CANCEL_RP_AGREEMENT_NOT_FOUND,
-  CANCEL_RP_LICENCE_NOT_FOUND,
-  CANCEL_RP_ALREADY_CANCELLED
-} from '../../uri.js'
 
 jest.mock('@defra-fish/connectors-lib')
-const getCancelRPURIs = () => [
-  CANCEL_RP_IDENTIFY.uri,
-  CANCEL_RP_DETAILS.uri,
-  CANCEL_RP_CONFIRM.uri,
-  CANCEL_RP_COMPLETE.uri,
-  CANCEL_RP_AGREEMENT_NOT_FOUND.uri,
-  CANCEL_RP_LICENCE_NOT_FOUND.uri,
-  CANCEL_RP_ALREADY_CANCELLED.uri
-]
+
 let TestUtils = null
 describe('Telesales route handlers', () => {
   // Start application before running the test case
@@ -59,37 +42,5 @@ describe('Telesales route handlers', () => {
       url: '/oidc/role-required'
     })
     expect(data.statusCode).toBe(200)
-  })
-})
-
-describe('cancellation route journey behaves as expected', () => {
-  beforeEach(jest.clearAllMocks)
-
-  it('adds the cancellation route journey if SHOW_CANCELLATION_JOURNEY is set to true', () => {
-    process.env.SHOW_CANCELLATION_JOURNEY = 'true'
-    jest.isolateModules(() => {
-      const telesalesRoutePaths = require('../telesales-routes.js').default.map(route => route.path)
-      expect(telesalesRoutePaths).toEqual(expect.arrayContaining(getCancelRPURIs()))
-    })
-  })
-
-  it('omits the cancellation route journey if SHOW_CANCELLATION_JOURNEY is set to false', () => {
-    process.env.SHOW_CANCELLATION_JOURNEY = 'false'
-    jest.isolateModules(() => {
-      const telesalesRoutes = require('../telesales-routes.js').default
-      const cancelRPURIs = getCancelRPURIs()
-      const cancelRPRoutes = telesalesRoutes.filter(route => cancelRPURIs.includes(route.path))
-      expect(cancelRPRoutes).toHaveLength(0)
-    })
-  })
-
-  it('omits the cancellation route journey if SHOW_CANCELLATION_JOURNEY is not present', () => {
-    delete process.env.SHOW_CANCELLATION_JOURNEY
-    jest.isolateModules(() => {
-      const telesalesRoutes = require('../telesales-routes.js').default
-      const cancelRPURIs = getCancelRPURIs()
-      const cancelRPRoutes = telesalesRoutes.filter(route => cancelRPURIs.includes(route.path))
-      expect(cancelRPRoutes).toHaveLength(0)
-    })
   })
 })
