+++ packages/gafl-webapp-service/src/routes/routes.js
@@ -32,6 +32,14 @@ import invalidLink from '../pages/renewals/renewal-inactive/route.js'
 import renewalStartDate from '../pages/renewals/renewal-start-date/route.js'
 import licenceNotFound from '../pages/renewals/licence-not-found/route.js'
 
+import cancelRPIdentify from '../pages/recurring-payments/cancel/identify/route.js'
+import cancelRPDetails from '../pages/recurring-payments/cancel/details/route.js'
+import cancelRPConfirm from '../pages/recurring-payments/cancel/confirm/route.js'
+import cancelRPComplete from '../pages/recurring-payments/cancel/complete/route.js'
+import cancelRPAgreementNotFound from '../pages/recurring-payments/cancel/agreement-not-found/route.js'
+import cancelRPLicenceNotFound from '../pages/recurring-payments/cancel/licence-not-found/route.js'
+import cancelRPAlreadyCancelled from '../pages/recurring-payments/cancel/already-cancelled/route.js'
+
 import staticAssets from './static-routes.js'
 import miscRoutes from './misc-routes.js'
 import telesalesRoutes from './telesales-routes.js'
@@ -75,6 +83,18 @@ const routes = [
   ...setUpRecurring,
   ...licenceNotFound
 ]
+
+if (process.env.SHOW_CANCELLATION_JOURNEY === 'true') {
+  routes.push(
+    ...cancelRPIdentify,
+    ...cancelRPDetails,
+    ...cancelRPConfirm,
+    ...cancelRPComplete,
+    ...cancelRPAgreementNotFound,
+    ...cancelRPAlreadyCancelled,
+    ...cancelRPLicenceNotFound
+  )
+}
 
 if (process.env.CHANNEL === 'telesales') {
   routes.push(...telesalesRoutes)
