+++ packages/gafl-webapp-service/src/routes/telesales-routes.js
@@ -1,13 +1,6 @@
 import { OIDC_SIGNIN, OIDC_ACCOUNT_DISABLED, OIDC_ROLE_REQUIRED, CONTROLLER } from '../uri.js'
 import { signIn } from '../handlers/oidc-handler.js'
 import journeyGoal from '../pages/journey-goal/route.js'
-import cancelRPIdentify from '../pages/recurring-payments/cancel/identify/route.js'
-import cancelRPDetails from '../pages/recurring-payments/cancel/details/route.js'
-import cancelRPConfirm from '../pages/recurring-payments/cancel/confirm/route.js'
-import cancelRPComplete from '../pages/recurring-payments/cancel/complete/route.js'
-import cancelRPAgreementNotFound from '../pages/recurring-payments/cancel/agreement-not-found/route.js'
-import cancelRPLicenceNotFound from '../pages/recurring-payments/cancel/licence-not-found/route.js'
-import cancelRPAlreadyCancelled from '../pages/recurring-payments/cancel/already-cancelled/route.js'
 
 const telesalesRoutes = [
   {
@@ -36,15 +29,4 @@ const telesalesRoutes = [
   ...journeyGoal
 ]
 
-if (process.env.SHOW_CANCELLATION_JOURNEY === 'true') {
-  telesalesRoutes.push(
-    ...cancelRPIdentify,
-    ...cancelRPDetails,
-    ...cancelRPConfirm,
-    ...cancelRPComplete,
-    ...cancelRPAgreementNotFound,
-    ...cancelRPAlreadyCancelled,
-    ...cancelRPLicenceNotFound
-  )
-}
 export default telesalesRoutes
